apply values from config

+++ AppDeployToolkit/AppDeployToolkitExtensions.ps1
@@ -5754,11 +5754,16 @@ function Initialize-NxtUninstallApplication {
 			if ($false -eq $uninstallKeyToHide.Is64Bit -and $true -eq $Is64Bit) {
 				$wowEntry = "\Wow6432Node"
 			}
-			if ($true -eq $uninstallKeyToHide.KeyNameIsDisplayName) {
-				[string]$currentKeyName = (Get-NxtInstalledApplication -UninstallKey $uninstallKeyToHide.KeyName -UninstallKeyIsDisplayName $true).UninstallSubkey
+			[PSCustomObject]$getInstalledApplicationSplatted = @{
+				UninstallKey = $uninstallKeyToHide.KeyName
+				UninstallKeyIsDisplayName = $uninstallKeyToHide.KeyNameIsDisplayName
+				UninstallKeyContainsWildCards = $uninstallKeyToHide.KeyNameContainsWildCards
+				DisplayNamesToExclude = $UninstallKeysToHide.DisplayNamesToExcludeFromHiding
 			}
-			else {
-				[string]$currentKeyName = $uninstallKeyToHide.KeyName
+			[string[]]$currentKeyName = (Get-NxtInstalledApplication @getInstalledApplicationSplatted).UninstallSubkey
+			if ($currentKeyName.Count -ne 1) {
+				Write-Log -Message "Did not find unique uninstall registry key with name [$($uninstallKeyToHide.KeyName)]. Skipped hiding the entry for this key." -Source ${CmdletName} -Severity 2
+				continue
 			}
 			if (Get-RegistryKey -Key "HKLM:\Software$wowEntry\Microsoft\Windows\CurrentVersion\Uninstall\$currentKeyName" -Value SystemComponent) {
 				Remove-RegistryKey -Key "HKLM:\Software$wowEntry\Microsoft\Windows\CurrentVersion\Uninstall\$currentKeyName" -Name 'SystemComponent'
